Enable the charging switch
The added code enables to switch charging on and off. It is kinda sloppy and a crappy solution, but it extends a bit the Battery mod battery life and will extend device battery lifetime since it won't charge over 80%.

It still needs buttons, toggles, notifications and to enter the main loop.